Fine tune load balancing
The default load balancing configuration does not check the status of
instances before sending them a request. If one of the instances goes
down, Eureka Server will only notice it after some time. This is
because clients need to renew their leases on the Eureka Server at
regular time intervals so that the server knows that they are still
alive. If one of the instances goes down, it will only be deregistered
after the Eureka Server has noticed that the instance did not renew its
lease within a given period. This means that within the period that the
dead instance has not been deregistered, the application will be
failing for requests sent to the dead instance. This problem has been
fixed.